requirements.txt
aiogram>=3.4.0
openai>=1.12.0
python-docx>=1.1.0
PyMuPDF>=1.24.0
markdown>=3.5.0
htmldocx>=0.0.6
